+++ changelog.md
@@ -23,6 +23,7 @@ This project attempts to adheres to [Semantic Versioning](http://semver.org/).
 - Functions.contant returns a function that always produces a constant return value
 - Functions.forMap returns a function that performs a map lookup returning an Option
 - Functions.forMapWithDefault returns a function that performs a map lookup returning a default value when it fails
+- Suppliers.memoize and Suppliers.weakMemoize provide call-by-need evaluation strategies
 
 ### Changed
 - Throwables and Function2 have been moved to fugue-guava package and the com.atlassian.fugue.deprecated package.